Mudar tipo da variável do resultado do fatorial

+++ algoritmos/fatorial/fatorial.c
@@ -1,7 +1,7 @@
 #include "fatorial.h"
 
-int fatorial (int parametro) {
-    int fatorial = 1;
+unsigned long long int fatorial (int parametro) {
+    unsigned long long int fatorial = 1;
     for (parametro; parametro > 1; parametro --) fatorial *= parametro;
     return fatorial;
 }
